requirements.txt
# Core Telegram Bot Framework
python-telegram-bot>=20.0,<21.0

# Bybit API Client
pybit>=5.0,<6.0

# HTTP Client and Async Support
aiohttp>=3.8.0,<4.0
httpx>=0.24.0,<1.0

# Environment Variables
python-dotenv>=1.0.0,<2.0

# AI/OpenAI Integration
openai>=1.0.0,<2.0

# Social Media APIs
praw>=7.7.0                    # Reddit API
tweepy>=4.14.0                 # Twitter API v2
google-api-python-client>=2.0  # YouTube Data API v3
discord.py>=2.3.0              # Discord API (optional)

# Enhanced NLP & Sentiment Analysis
textblob>=0.17.1               # Basic sentiment analysis
vaderSentiment>=3.3.2          # Social media optimized sentiment
# transformers>=4.30.0         # Advanced NLP models (optional - large dependency)
# nltk>=3.8.1                  # Natural language toolkit (optional)

# Image Processing for GGShot Feature
pillow>=9.0.0,<11.0            # Image processing and enhancement
# opencv-python>=4.8.0,<5.0    # Advanced image processing (optional - uncomment if needed)

# Web Scraping & Data Processing
beautifulsoup4>=4.12.0         # Web scraping utilities
# newspaper3k>=0.2.8           # News article processing (optional)

# Data Processing and Math
pandas>=2.0.0,<3.0
numpy>=1.24.0,<2.0

# Async Support (usually included with Python 3.7+)
asyncio-mqtt>=0.16.0,<1.0

# Additional Utilities
requests>=2.28.0,<3.0
urllib3>=1.26.0,<2.0

# Time and Date Handling
python-dateutil>=2.8.0,<3.0

# Decimal Precision (usually built-in)
# decimal - built into Python

# Logging (built-in but enhanced formatters)
colorlog>=6.7.0,<7.0

# Development and Testing (optional)
pytest>=7.0.0,<8.0
pytest-asyncio>=0.21.0,<1.0

# Memory Profiling (optional - for debugging)
memory-profiler>=0.60.0,<1.0
psutil>=5.9.0,<6.0